TPs/TP2/requirements.txt
aiohttp==3.10.10 
Pillow==10.1.0